+++ Sources/PrintingKit/Pdf/Pdf.swift
@@ -1,0 +1,12 @@
+//
+//  Pdf.swift
+//  PrintingKit
+//
+//  Created by Daniel Saidi on 2023-08-21.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/// This namespace contains PDF-specific types.
+public struct Pdf {}